fixed playing suicidally and placing stones on occupied points

+++ go_obj.py
@@ -7,11 +7,11 @@ class Group(object):
         self.size = size
         self.color = color
         self.points = [coords]
-    
+
     # merges points from another group into this group.  other group should be subsequently deleted.
     def merge(self, other):
         self.points += other.points
-    
+
     #returns all points adjacent to a group
     def get_adjacent_points(self):
         direction_vectors = ((1, 0), (0, 1), (-1, 0), (0, -1))
@@ -23,14 +23,14 @@ class Group(object):
                     if check[0] >= 0 and check[0] < self.size and check[1] >= 0 and check[1] <self.size:
                         points.append(check)
         return points
-    
+
 class Board(object):
 
     str_translator = {None: " +  ", "Black": " B  ", "White": " W  "}
 
     cols = ['A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'J',
         'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T']
-    
+
     def __init__(self, size, groups):
         self.size = size
         self.groups = groups
@@ -43,7 +43,7 @@ class Board(object):
         for group in groups:
             for point in group.points:
                 self.matrix[point[0]][point[1]] = group.color
-    
+
     # formats a Board for printing
     def __str__(self):
         formatted = '\n    '
@@ -61,14 +61,26 @@ class Board(object):
         return formatted
 
     # returns a list of groups that are dead
-    def compute_kills(self):
+    def compute_kills(self, move):
         kills = []
+        move_is_dead = False
         for group in self.groups:
-            liberties = self.find_liberties(group)
-            if len(liberties) is 0:
-                kills.append(group)
-        return kills
-    
+            if len(self.find_liberties(group)) is 0:
+                if group is move:
+                    move_is_dead = True
+                else:
+                    kills.append(group)
+        if move_is_dead:
+            liberty_checks = move.get_adjacent_points()
+            for point in liberty_checks:
+                for group in kills:
+                    if point in group.points:
+                        move_is_dead = False
+        if move_is_dead:
+            return [move]
+        else:
+            return kills
+
     # returns a list of liberties for a group
     def find_liberties(self, group):
         liberties = []
@@ -76,9 +88,9 @@ class Board(object):
             if self.matrix[point[0]][point[1]] is None:
                 liberties.append(point)
         return liberties
-    
+
 class Game(object):
-    
+
     score = {"Black": 0, "White": 6.5}
 
     teams = ('White', 'Black')
@@ -92,15 +104,18 @@ class Game(object):
     def __init__(self, size):
         self.size = size
         self.board = Board(size, self.groups)
-        self.cols = self.cols[:size-1]
+        self.last_board = self.board
+        self.cols = self.cols[:size]
         self.rows = range(1, size + 1)
-    
+
     # performs all actions relevant to a turn.  returns true if the game is over
     def next_turn(self):
         self.turn_number += 1
+        self.last_board = self.board
         color = self.teams[self.turn_number%2]
         print(self.board)
         while True:
+            reset_loop = False
             print(f"\n{color}'s turn, where would you like to play?")
             couplet = input("\n> ").upper()
             if couplet == "PASS":
@@ -118,12 +133,19 @@ class Game(object):
                     print("\nEnding Game...")
                     print(self.board)
                     return True
-            if len(couplet) is not 2:
+            if len(couplet) is not 2 and len(couplet) is not 3:
                 print('\nPlease specify your move in the form "A9" or "pass"')
                 continue
-            if couplet[0] in self.cols and int(couplet[1]) in self.rows:
+            if couplet[0] in self.cols and int(couplet[1:]) in self.rows:
                 self.last_turn_passed = False
                 coords = self.couplet_to_coords(couplet)
+                for group in self.groups:
+                    if coords in group.points:
+                        print("There is already a stone there")
+                        reset_loop = True
+                        break
+                if reset_loop:
+                    continue
                 break
             else:
                 print('\nOut of Bounds')
@@ -136,8 +158,14 @@ class Game(object):
                         self.groups.remove(group)
         self.groups.append(move)
         self.board = Board(self.size, self.groups)
-        kills = self.board.compute_kills()
+        kills = self.board.compute_kills(move)
         if len(kills) is not 0:
+            if kills[0] is move:
+                print("That is Suicidal")
+                self.turn_number -= 1
+                self.board = self.last_board
+                return self.next_turn()
+
             for kill in kills:
                 self.score[color] += len(kill.points)
                 self.groups.remove(kill)
@@ -147,8 +175,8 @@ class Game(object):
 
     # returns a 2d tuple containing the coords associated with a couplet ex: "A0" -> (0, 0)
     def couplet_to_coords(self, couplet):
-        return (int(couplet[1])-1, self.cols.index(couplet[0]))
-    
+        return (int(couplet[1:])-1, self.cols.index(couplet[0]))
+
     #prints the score, ugly but easier than making an entire score class with a __str__ method
     def print_score(self):
         print("\nWhite: {}".format(self.score["White"]))
